running into an issue of having data move from one component to another
