+++ src/components/UserInputs.js
@@ -1,17 +1,15 @@
-import Results from "./Results.js";
+
 import { useEffect, useState } from 'react'
-import axios from 'axios'
+
+import { Link } from "react-router-dom";
 
 
 const UserInputs = () => {
 
-    const [userChoice, setUserChoice] = useState("placeholder");
-    const [jokes, setJokes] = useState([]);
-    const [apiError, setApiError] = useState(false);
-    const [searchTerm, setSearchTerm] = useState('');
-    const [photos, setPhotos] = useState([]);
-    const [submit, setSubmit] = useState(false);
     const [userLimitChoice, setuserLimitChoice] = useState(10);
+    const [searchTerm, setSearchTerm] = useState('');
+    const [userChoice, setUserChoice] = useState("placeholder");
+    const [submit, setSubmit] = useState(false);
     const [usersDadJokes, setUsersDadJokes] = useState([]);
 
 
@@ -20,73 +18,16 @@ const UserInputs = () => {
         setuserLimitChoice(userLimitChoice - usersDadJokes.length);
         setSearchTerm(userChoice);
         setSubmit(true);
-    }
-
-
-    const handleUserChoice = (e) => {
+      }
+    
+    
+      const handleUserChoice = (e) => {
         setUserChoice(e.target.value);
-    }
-
-
-    // const handleButtonClick = (e) => {
-    //     setRandomButton(true);
-    // }
-
-    const apiKey = `34_FRr4gH3efbjKeNMjRmPjTM8phiy64ND24X1GElr8`
-
-    useEffect(() => {
-        axios({
-            url: `https://api.unsplash.com/search/photos`,
-            dataResponse: `json`,
-            method: `GET`,
-            params: {
-                client_id: apiKey,
-                query: searchTerm,
-                per_page: 10,
-            }
-        }).then((response) => {
-            setPhotos(response.data.results);
-        }).catch((error) => {
-            setApiError(error)
-        })
-    }, [searchTerm]);
-
-
-    const randomNumber = Math.floor(Math.random() * 64);
-
-
+      }
     
-
-    useEffect(() => {
-        if(submit){
-            axios({
-                url: `https://icanhazdadjoke.com/search`,
-                dataResponse: `json`,
-                method: `GET`,
-                headers: {
-                    "Accept": "application/json"
-                },
-                params: {
-                    limit: userLimitChoice,
-                    page: randomNumber,
-                    total_jokes: 100
-                }
-            }).then((response) => {
-                setJokes(response.data.results)
-            }).catch((error) => {
-                setApiError(error)
-            })
-        }
-    }, [submit])
-
-    
-
-    
-
     return (
         <>
             <div>
-                <button>Need a Laugh?</button>
                 <form
                     onSubmit={handleSubmit}
                 >
@@ -101,10 +42,11 @@ const UserInputs = () => {
                         <option value="cats">Cats</option>
                         <option value="bunnies">Bunnies</option>
                     </select>
-                    <button type="submit">Submit</button>
+                    <Link to='/results' >
+                        <button type="submit">Submit</button>
+                    </Link>
                 </form>
             </div>
-            <Results photos={photos} jokes={jokes} />
         </>
     )
